src/main/Position.java
/**
 * This class represents the position of a robot - its coordinates (x, y) and direction.
 */

package main;

public class Position {
    private int x;
    private int y;
    private char direction;

    public Position(int x, int y, char direction) {
        this.x = x;
        this.y = y;
        this.direction = direction;
    }

    public int getX() {
        return x;
    }

    public int getY() {
        return y;
    }

    public char getDirection() {
        return direction;
    }
}
